change(ios): Centralizes kbd download error alerts
